Project Oracle/requirements.txt
streamlit
PyGithub
python-dotenv
requests